requirements.txt
streamlit
folium
streamlit_folium
Pillow